Moving stuff around and added the listener
