+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".WifiAccessActivity"
+            android:name=".activity.WifiAccessActivity"
             android:label="@string/title_activity_wifi_access"
             android:theme="@style/AppTheme">
             <intent-filter>
