+++ app/src/main/java/com/android/andrew/wifi_selector/activity/AppCompatPreferenceActivity.java
@@ -1,4 +1,4 @@
-package com.android.andrew.wifi_selector;
+package com.android.andrew.wifi_selector.activity;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
