+++ app/src/main/res/layout/activity_wifi_access.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".WifiAccessActivity">
+    tools:context=".activity.WifiAccessActivity">
 
     <include layout="@layout/content_wifi_access" />
