+++ app/src/main/res/layout/content_wifi_access.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".WifiAccessActivity"
+    tools:context=".activity.WifiAccessActivity"
     tools:showIn="@layout/activity_wifi_access">
 
     <ListView
