+++ app/src/main/res/xml/pref_headers.xml
@@ -3,17 +3,17 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="com.android.andrew.wifi_selector.SettingsActivity$GeneralPreferenceFragment"
+        android:fragment="com.android.andrew.wifi_selector.activity.SettingsActivity$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="com.android.andrew.wifi_selector.SettingsActivity$NotificationPreferenceFragment"
+        android:fragment="com.android.andrew.wifi_selector.activity.SettingsActivity$NotificationPreferenceFragment"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/pref_header_notifications" />
 
     <header
-        android:fragment="com.android.andrew.wifi_selector.SettingsActivity$DataSyncPreferenceFragment"
+        android:fragment="com.android.andrew.wifi_selector.activity.SettingsActivity$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
